clean up parser file and function
